A better fix for appveyor overflow (last one was silly).

+++ plasmapy/physics/transport.py
@@ -2125,7 +2125,7 @@ def _nondim_tc_e_ji_held(hall, Z, field_orientation):
 
     allowed_Z = [1, 2, 'arbitrary']
     Z_idx = _check_Z(allowed_Z, Z)
-    r = np.abs(Z * hall)
+    r = float(np.abs(Z * hall))
 
     def f_kappa_par_e(Z):
         numerator = 13.5 * Z ** 2 + 54.4 * Z + 25.2
@@ -2227,7 +2227,7 @@ def _nondim_tc_e_ji_held(hall, Z, field_orientation):
 #                       k_2[Z_idx] * r ** (5 / 3) +
 #                       k_1[Z_idx] * r +
 #                       k_0[Z_idx])
-        term1 = float(r ** 3)
+        term1 = r ** 3
         term2 = k_4[Z_idx] * r ** (7 / 3)
         term3 = k_3[Z_idx] * r ** 2
         term4 = k_2[Z_idx] * r ** (5 / 3)
